requirements.txt
fastapi==0.110.1
pydantic==2.7.0
SQLAlchemy==2.0.29
uvicorn==0.29.0
asyncpg==0.29.0
python-multipart==0.0.9
aiofiles==23.2.1
python-dotenv==1.0.1